Order api created

+++ src/main/java/com/pratik/model/Order.java
@@ -18,29 +18,20 @@ public class Order {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id;
-
     @ManyToOne
     private User customer;
-
     @JsonIgnore
     @ManyToOne
     private Restaurant restaurant;
-
     private Long totatAmount;
-
     private String orderStatus;
-
     private Date createdAt;
-
     @ManyToOne
     private Address deliveryAddress;
-
     @OneToMany
     private List<OrderItem> items;
-
     private int totalItem;
-
-    private int totalPrice;
+    private Long totalPrice;
 
 //    private Payment paymentMethod;
 }
